perf(core): Skip worker threads for lightweight file processing
When compress mode is off (the default), file processing only does
lightweight regex/string operations (truncateBase64, removeComments,
removeEmptyLines, trim). Worker thread overhead (pool creation,
structured clone of all file content, message passing, WASM module
loading) far exceeds the computation cost.

- CHANGE: Use main-thread promisePool instead of worker threads when
  compress is disabled; extract shared promisePool utility from
  fileCollect for reuse
- WHY: Benchmarks show -6% (Ubuntu/macOS) to -17% (Windows) improvement
  by avoiding worker overhead for simple string operations
- CONSTRAINT: Compress mode still requires workers because tree-sitter
  WASM parsing is CPU-heavy per file
